Split resemblesUrl in lib/url.
